requirements.txt
blinker==1.9.0
click==8.3.1
et_xmlfile==2.0.0
Flask==2.3.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==8.1.0
numpy==1.23.5
openpyxl==3.0.10
pandas==1.5.3
protobuf==4.21.12
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.3
Werkzeug==3.1.5
dotenv==0.0.5